Move CompletionCallback to sdk-common module
