+++ MapboxSearch/base/src/main/java/com/mapbox/search/base/StubCompletionCallback.kt
@@ -1,0 +1,9 @@
+package com.mapbox.search.base
+
+import com.mapbox.search.common.CompletionCallback
+
+class StubCompletionCallback<T> : CompletionCallback<T> {
+    override fun onComplete(result: T) {}
+
+    override fun onError(e: Exception) {}
+}
