+++ MapboxSearch/sdk-common/api/sdk-common.api
@@ -18,6 +18,11 @@ public final class com/mapbox/search/common/BuildConfig {
 	public fun <init> ()V
 }
 
+public abstract interface class com/mapbox/search/common/CompletionCallback {
+	public abstract fun onComplete (Ljava/lang/Object;)V
+	public abstract fun onError (Ljava/lang/Exception;)V
+}
+
 public abstract interface class com/mapbox/search/common/DistanceCalculator {
 	public static final field Companion Lcom/mapbox/search/common/DistanceCalculator$Companion;
 	public abstract fun distance (Lcom/mapbox/geojson/Point;Lcom/mapbox/geojson/Point;)D
