+++ MapboxSearch/sdk/src/main/java/com/mapbox/search/IndexableDataProvidersRegistry.kt
@@ -2,6 +2,7 @@ package com.mapbox.search
 
 import com.mapbox.search.base.core.CoreSearchEngineInterface
 import com.mapbox.search.common.AsyncOperationTask
+import com.mapbox.search.common.CompletionCallback
 import com.mapbox.search.record.IndexableDataProvider
 import com.mapbox.search.record.IndexableRecord
 import java.util.concurrent.Executor
