+++ MapboxSearch/sdk/src/main/java/com/mapbox/search/SearchEngine.kt
@@ -1,7 +1,9 @@
 package com.mapbox.search
 
 import com.mapbox.search.analytics.AnalyticsService
+import com.mapbox.search.base.StubCompletionCallback
 import com.mapbox.search.common.AsyncOperationTask
+import com.mapbox.search.common.CompletionCallback
 import com.mapbox.search.common.concurrent.SearchSdkMainThreadWorker
 import com.mapbox.search.record.IndexableDataProvider
 import com.mapbox.search.record.IndexableRecord
