+++ MapboxSearch/sdk/src/main/java/com/mapbox/search/record/HistoryDataProvider.kt
@@ -1,6 +1,5 @@
 package com.mapbox.search.record
 
-import com.mapbox.search.CompletionCallback
 import com.mapbox.search.SearchResultMetadata
 import com.mapbox.search.base.record.BaseIndexableRecord
 import com.mapbox.search.base.record.SearchHistoryService
@@ -10,6 +9,7 @@ import com.mapbox.search.base.utils.LocalTimeProvider
 import com.mapbox.search.base.utils.TimeProvider
 import com.mapbox.search.base.utils.extension.mapToPlatform
 import com.mapbox.search.common.AsyncOperationTask
+import com.mapbox.search.common.CompletionCallback
 import com.mapbox.search.result.mapToPlatform
 import java.util.PriorityQueue
 import java.util.concurrent.Executor
