+++ MapboxSearch/sdk/src/main/java/com/mapbox/search/record/IndexableDataProvider.kt
@@ -1,7 +1,7 @@
 package com.mapbox.search.record
 
-import com.mapbox.search.CompletionCallback
 import com.mapbox.search.common.AsyncOperationTask
+import com.mapbox.search.common.CompletionCallback
 import com.mapbox.search.common.concurrent.SearchSdkMainThreadWorker
 import java.util.concurrent.Executor
